Ikke send doble events ved sidevisning
