+++ src/Forside/Forside.tsx
@@ -1,57 +1,14 @@
-import * as React from 'react';
-import { useEffect } from 'react';
-import './Forside.less';
-import ManglerRettigheterIAltinnSide from '../FeilSider/ManglerRettigheterIAltinnSide/ManglerRettigheterIAltinnSide';
-import Innloggingsside from '../Innloggingsside/Innloggingsside';
-import { RestSykefraværshistorikk } from '../api/sykefraværshistorikk';
-import { RestStatus } from '../api/api-utils';
-import { RestAltinnOrganisasjoner } from '../api/altinnorganisasjon-api';
+import { default as React, FunctionComponent } from 'react';
 import { useOrgnr } from '../utils/orgnr-hook';
-import { useSendEvent } from '../amplitude/amplitude';
+import { useSendSidevisningEvent } from '../amplitude/amplitude';
 
-interface Props {
-    restSykefraværshistorikk: RestSykefraværshistorikk;
-    restOrganisasjonerMedStatistikk: RestAltinnOrganisasjoner;
-}
-
-const Forside: React.FunctionComponent<Props> = ({
-    restSykefraværshistorikk,
-    restOrganisasjonerMedStatistikk,
-    children,
-}) => {
+export const Forside: FunctionComponent = (props) => {
     const orgnr = useOrgnr();
-    const status = restSykefraværshistorikk.status;
-    const sendEvent = useSendEvent();
+    useSendSidevisningEvent('forside', orgnr);
 
-    useEffect(() => {
-        if (status === RestStatus.Suksess) {
-            sendEvent('forside', 'vist');
-        } else if (status === RestStatus.IngenTilgang) {
-            sendEvent('forside', 'vist', {
-                ingenTilgang: true,
-            });
-        }
-    }, [orgnr, status, sendEvent]);
-
-    switch (status) {
-        case RestStatus.IngenTilgang: {
-            return (
-                <ManglerRettigheterIAltinnSide
-                    restOrganisasjonerMedStatistikk={restOrganisasjonerMedStatistikk}
-                />
-            );
-        }
-        case RestStatus.IkkeInnlogget: {
-            return <Innloggingsside />;
-        }
-        default: {
-            return (
-                <div className="forside__wrapper">
-                    <div className="forside">{children}</div>
-                </div>
-            );
-        }
-    }
+    return (
+        <div className="forside__wrapper">
+            <div className="forside">{props.children}</div>
+        </div>
+    );
 };
-
-export default Forside;
